Fix UI Stretch, NPC Interaction, Extra Sounds and Arrows, Game Icon, Updated Controls and Credits
Fixed UI stretching by changing all Canvases to "Scale with Screen Size." Added extra NPC interaction by having the player press T when interacting with the Tourist to cause him to react, then go through the door to level 2. Also added extra sounds and arrows over the door to the tourist when all spirit orbs are collected to make it clearer to the player. Also added a game icon, as opposed to using the team logo there. Also updated the controls with the NPC interaction and updated the credits screen with the free sounds and music used and team updates.