Attack and Fan Animations
Attack and Fan Animations Playing Correctly. Found and fixed slight attack bug, where it would not trigger at the start of the levels, because isRight and isLeft on player movement script is not set until you move. Fixed to initialize as true or false based on open scene in movement script.